Move common objects and methods to perun beans module
 - move all serialiezers and deserializers from rpc-lib and rpc to
   perun-beans
 - move all registrar models to perun-beans
 - move some core objects (PerunRequest, AuthzRoles, PerunPrincipal),
   because it is needed to have them in common module
 - move methods getPerunConfigFiles* from Utils to BeanUtils
 - copy method notNull from Utils to BeansUtils (not move because too
   simple method and too much uses in perun-core)
 - change all affected pom.xml to add or remove some dependencies
 - create new application context for beans-utils and move properties
   from perun-core here
 - fix imports and BeansUtils where needed
