+++ perun-beans/src/main/java/cz/metacentrum/perun/registrar/model/ApplicationMail.java
