+++ perun-beans/src/main/java/cz/metacentrum/perun/rpc/deserializer/Deserializer.java
@@ -1,7 +1,7 @@
 package cz.metacentrum.perun.rpc.deserializer;
 
 import cz.metacentrum.perun.core.api.PerunBean;
-import cz.metacentrum.perun.rpc.RpcException;
+import cz.metacentrum.perun.core.api.exceptions.RpcException;
 
 import javax.servlet.http.HttpServletRequest;
 import java.util.List;
