+++ perun-beans/src/main/java/cz/metacentrum/perun/rpc/deserializer/UrlDeserializer.java
@@ -6,7 +6,7 @@ import java.util.List;
 
 import javax.servlet.http.HttpServletRequest;
 
-import cz.metacentrum.perun.rpc.RpcException;
+import cz.metacentrum.perun.core.api.exceptions.RpcException;
 
 /**
  * Deserializer for URL data format.
