+++ perun-beans/src/main/java/cz/metacentrum/perun/rpc/serializer/JsonSerializerJSONP.java
@@ -4,7 +4,7 @@ import cz.metacentrum.perun.cabinet.model.Authorship;
 import cz.metacentrum.perun.core.api.*;
 import cz.metacentrum.perun.core.api.exceptions.PerunException;
 import cz.metacentrum.perun.core.api.exceptions.rt.PerunRuntimeException;
-import cz.metacentrum.perun.rpc.RpcException;
+import cz.metacentrum.perun.core.api.exceptions.RpcException;
 import org.codehaus.jackson.JsonEncoding;
 import org.codehaus.jackson.JsonFactory;
 import org.codehaus.jackson.JsonGenerator;
