+++ perun-core/src/main/resources/perun-beans-test.xml
@@ -7,9 +7,10 @@
 http://www.springframework.org/schema/tx http://www.springframework.org/schema/tx/spring-tx.xsd
 http://www.springframework.org/schema/aop http://www.springframework.org/schema/aop/spring-aop.xsd http://www.springframework.org/schema/jdbc http://www.springframework.org/schema/jdbc/spring-jdbc.xsd">
 
+	<import resource="classpath:perun-beans-appcontext.xml"/>
     <import resource="classpath:perun-transaction-manager.xml"/>
     <import resource="perun-datasources-test.xml"/>
-
+	
     <aop:config>
         <aop:advisor advice-ref="txAdviceReadOnly" pointcut="execution(* cz.metacentrum.perun.core.entry.ServicesManagerEntry.getHierarchicalData(..))"/>
         <aop:advisor advice-ref="txAdviceReadOnly" pointcut="execution(* cz.metacentrum.perun.core.entry.ServicesManagerEntry.getDataWithGroups(..))"/>
@@ -62,14 +63,6 @@ http://www.springframework.org/schema/aop http://www.springframework.org/schema/
 
     <bean class="cz.metacentrum.perun.core.blImpl.AuthzResolverBlImpl" factory-method="setPerunBlImpl" scope="singleton">
         <constructor-arg ref="perun" />
-    </bean>
-
-    <!-- FIXME  hack, hackiti hack - inject properties bean into static class Utils
-                 (creates another bean, that is never used)
-                 The point is to stop Utils from complaining about non-existent /etc/perun/perun.properties.
-    -->
-    <bean id="utilsProperties" class="cz.metacentrum.perun.core.impl.Utils" factory-method="setProperties" scope="singleton">
-        <constructor-arg ref="coreProperties" />
     </bean>
 
     <!-- Perun implementation -->
@@ -278,29 +271,6 @@ http://www.springframework.org/schema/aop http://www.springframework.org/schema/
     <bean id="auditerConsumer" class="cz.metacentrum.perun.core.impl.AuditerConsumer" scope="prototype">
         <constructor-arg type="java.lang.String" value="consumerForTest" />
         <constructor-arg ref="dataSource" />
-    </bean>
-
-    <!-- Properties Bean for perun-core -->
-    <bean id="coreProperties" class="org.springframework.beans.factory.config.PropertiesFactoryBean">
-        <property name="ignoreResourceNotFound" value="true" />
-        <property name="properties">
-            <props>
-                <prop key="perun.admins">perunTests, perunController, perunEngine, perunRegistrar, perunSynchronizer, perunCabinet</prop>
-                <prop key="perun.engine.principals">perunv3/engine@META</prop>
-                <prop key="perun.registrar.principals">perunv3-registrar@META</prop>
-                <prop key="perun.notification.principals">perunNotifications</prop>
-                <prop key="perun.rpc.principal">perunRpc</prop>
-                <prop key="perun.db.type">hsqldb</prop>
-                <prop key="perun.group.synchronization.interval">1</prop>
-                <prop key="perun.group.synchronization.timeout">10</prop>
-                <prop key="perun.rpc.powerusers">perunv3-registrar@META</prop>
-                <prop key="perun.perun.db.name">v3dev</prop>
-                <prop key="perun.rt.url">https://rt3.cesnet.cz/rt/REST/1.0/ticket/new</prop>
-                <prop key="perun.rt.serviceuser.username">perunv3-rt</prop>
-                <prop key="perun.rt.serviceuser.password">password</prop>
-                <prop key="perun.passwordManager.program">/usr/local/bin/perun.passwordManager</prop>
-            </props>
-        </property>
     </bean>
 
     <import resource="classpath:perun-core-trace-log.xml"/>
