+++ perun-registrar-lib/pom.xml
@@ -80,6 +80,12 @@
 
 		<dependency>
 			<groupId>cz.metacentrum.perun</groupId>
+			<artifactId>perun-beans</artifactId>
+			<version>${project.version}</version>
+		</dependency>
+
+		<dependency>
+			<groupId>cz.metacentrum.perun</groupId>
 			<artifactId>perun-core</artifactId>
 			<version>${project.version}</version>
 		</dependency>
