+++ perun-rpc-lib/src/main/java/cz/metacentrum/perun/rpclib/api/RpcCaller.java
@@ -3,6 +3,7 @@ package cz.metacentrum.perun.rpclib.api;
 import java.util.Map;
 
 import cz.metacentrum.perun.core.api.exceptions.PerunException;
+import cz.metacentrum.perun.rpclib.api.Deserializer;
 
 public interface RpcCaller {
 	public Deserializer call(String managerName, String methodName) throws PerunException;
