+++ perun-rpc/src/main/java/cz/metacentrum/perun/rpc/PerunManager.java
@@ -3,6 +3,7 @@ package cz.metacentrum.perun.rpc;
 import cz.metacentrum.perun.rpc.methods.*;
 import cz.metacentrum.perun.core.api.exceptions.PerunException;
 import cz.metacentrum.perun.rpc.deserializer.Deserializer;
+import cz.metacentrum.perun.core.api.exceptions.RpcException;
 
 public enum PerunManager {
